Inline ckbtc_minter_batching
